-Wall
git-svn-id: svn+ssh://oberon/svn/OMiSCID-Dev@992 c4bd1236-c4f7-0310-a2fe-a60683d9aa4a

+++ ServiceControl/DnsSdService.cpp
@@ -417,7 +417,7 @@ bool RegisterService::Register()
 	snprintf( ProtocolAndTransport, ProtocolAndTransportField+1, "%s.%s", Protocol, Transport );
 
 	err = DNSServiceRegister( &DnsSdConnection, 0, 0, Name, ProtocolAndTransport, Domain, NULL, (uint16_t)htons(Port),
-		Properties.GetTXTRecordLength(), Properties.ExportTXTRecord(), DnsRegisterReply, (void*)this );
+		(uint16_t)Properties.GetTXTRecordLength(), Properties.ExportTXTRecord(), DnsRegisterReply, (void*)this );
 
 //	err = DNSServiceRegister( &DnsSdConnection, 0, 0, Name, ProtocolAndTransport, Domain, NULL, (uint16_t)htons(Port),
 //		0, "", DnsRegisterReply, (void*)this );
